Add unit tests for HTML chunker
