+++ native/src/test/resources/html-chunker-test/expected/sample1_500_50.txt
@@ -1,7 +1,7 @@
 500 50
 
 --- Chunk 1 ---
-Metadata: {index=0, id=71}
+Metadata: {index=0, id=70}
 <!DOCTYPE html>
 <html>
 <head>
@@ -14,7 +14,7 @@ Metadata: {index=0, id=71}
     <h2>Introduction</h2>
 
 --- Chunk 2 ---
-Metadata: {h1=Sample HTML Document, index=1, id=72}
+Metadata: {h1=Sample HTML Document, index=1, id=71}
 
     <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.</p>
 
@@ -24,7 +24,7 @@ Metadata: {h1=Sample HTML Document, index=1, id=72}
     
 
 --- Chunk 3 ---
-Metadata: {h1=Sample HTML Document, index=2, h2=Main Content, id=79}
+Metadata: {h1=Sample HTML Document, index=2, h2=Main Content, id=78}
 <h2>Main Content</h2>
 
     <h3>Section 1: Getting Started</h3>
@@ -34,7 +34,7 @@ Metadata: {h1=Sample HTML Document, index=2, h2=Main Content, id=79}
     <h3>Section 2: Advanced Topics</h3>
 
 --- Chunk 4 ---
-Metadata: {h1=Sample HTML Document, index=3, h2=Main Content, h3=Section 2: Advanced Topics, id=85}
+Metadata: {h1=Sample HTML Document, index=3, h2=Main Content, h3=Section 2: Advanced Topics, id=84}
 <h3>Section 2: Advanced Topics</h3>
     <p>Neque porro quisquam est, qui dolorem ipsum quia dolor sit amet, consectetur, adipisci velit, sed quia non numquam eius modi tempora incidunt ut labore et dolore magnam aliquam quaerat voluptatem.</p>
 
@@ -44,7 +44,7 @@ Metadata: {h1=Sample HTML Document, index=3, h2=Main Content, h3=Section 2: Adva
     <h4>Subsection 2.2</h4>
 
 --- Chunk 5 ---
-Metadata: {h1=Sample HTML Document, index=4, id=91}
+Metadata: {h1=Sample HTML Document, index=4, id=90}
 <h4>Subsection 2.2</h4>
     <p>Quis autem vel eum iure reprehenderit qui in ea voluptate velit esse quam nihil molestiae consequatur, vel illum qui dolorem eum fugiat quo voluptas nulla pariatur?</p>
 
@@ -54,7 +54,7 @@ Metadata: {h1=Sample HTML Document, index=4, id=91}
     <h3>Final Thoughts</h3>
 
 --- Chunk 6 ---
-Metadata: {h1=Sample HTML Document, index=5, h2=Conclusion, h3=Final Thoughts, id=93}
+Metadata: {h1=Sample HTML Document, index=5, h2=Conclusion, h3=Final Thoughts, id=92}
 <h3>Final Thoughts</h3>
     <p>Similique sunt in culpa qui officia deserunt mollitia animi, id est laborum et dolorum fuga. Et harum quidem rerum facilis est et expedita distinctio.</p>
     <p>Nam libero tempore, cum soluta nobis est eligendi optio cumque nihil impedit quo minus id quod maxime placeat facere possimus, omnis voluptas assumenda est, omnis dolor repellendus.</p>
