+++ native/src/test/resources/markdown-chunker-test/expected/sample3_500_0.txt
@@ -1,7 +1,7 @@
 500 0
 
 --- Chunk 1 ---
-Metadata: {index=0, id=3596}
+Metadata: {index=0, id=12}
 # Short Document
 
 This is a simple markdown document with minimal content for testing edge cases.
