+++ native/src/test/resources/markdown-chunker-test/expected/sample4_500_0.txt
@@ -1,48 +1,44 @@
 500 0
 
 --- Chunk 1 ---
-Metadata: {index=0, id=3984}
+Metadata: {index=0, id=449}
 # Document Title
 
 This is the introduction to the document with some basic content.
 
 ## Long Section Header
 
-
+This section contains a very long piece of text that should span multiple chunks when processed with the default chunk size of 500 characters. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure
 
 --- Chunk 2 ---
-Metadata: {header=Long Section Header, index=1, id=3986}
-This section contains a very long piece of text that should span multiple chunks when processed with the default chunk size of 500 characters. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint
+Metadata: {header=Long Section Header, index=1, id=527}
+ dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
+
+Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt explicabo. Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut fugit, sed quia
 
 --- Chunk 3 ---
-Metadata: {header=Long Section Header, index=2, id=3989}
- occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
-
-
+Metadata: {header=Long Section Header, index=2, id=567}
+ consequuntur magni dolores eos qui ratione voluptatem sequi nesciunt. Neque porro quisquam est, qui dolorem ipsum quia dolor sit amet, consectetur, adipisci velit, sed quia non numquam eius modi tempora incidunt ut labore et dolore magnam aliquam quaerat voluptatem.
 
 --- Chunk 4 ---
-Metadata: {header=Long Section Header, index=3, id=3991}
-Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt explicabo. Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut fugit, sed quia consequuntur magni dolores eos qui ratione voluptatem sequi nesciunt. Neque porro quisquam est, qui dolorem ipsum quia dolor sit amet, consectetur, adipisci velit, sed quia non numquam eius modi
+Metadata: {index=3, id=570}
 
---- Chunk 5 ---
-Metadata: {header=Long Section Header, index=4, id=3993}
- tempora incidunt ut labore et dolore magnam aliquam quaerat voluptatem.
 
 At vero eos et accusamus et iusto odio dignissimos ducimus qui blanditiis praesentium voluptatum deleniti atque corrupti quos dolores et quas molestias excepturi sint occaecati cupiditate non provident, similique sunt in culpa qui officia deserunt mollitia animi, id est laborum et dolorum fuga. Et harum quidem rerum facilis est et expedita distinctio.
 
 ### Subsection
 
 
---- Chunk 6 ---
-Metadata: {index=5, id=3995}
+--- Chunk 5 ---
+Metadata: {index=4, id=573}
 
 This is a subsection under the long section. It also contains substantial content that may span chunks. Nam libero tempore, cum soluta nobis est eligendi optio cumque nihil impedit quo minus id quod maxime placeat facere possimus, omnis voluptas assumenda est, omnis dolor repellendus. Temporibus autem quibusdam et aut officiis debitis aut rerum necessitatibus saepe eveniet ut et voluptates repudiandae sint et molestiae non recusandae.
 
 ## Another Section
 
 
---- Chunk 7 ---
-Metadata: {header=Another Section, index=6, id=3997}
+--- Chunk 6 ---
+Metadata: {header=Another Section, index=5, id=575}
 
 This is a shorter section to test the transition between different header levels and content lengths.
