+++ native/src/test/resources/markdown-chunker-test/expected/sample4_500_50.txt
@@ -1,7 +1,7 @@
 500 50
 
 --- Chunk 1 ---
-Metadata: {index=0, id=578}
+Metadata: {index=0, id=449}
 # Document Title
 
 This is the introduction to the document with some basic content.
@@ -11,17 +11,17 @@ This is the introduction to the document with some basic content.
 This section contains a very long piece of text that should span multiple chunks when processed with the default chunk size of 500 characters. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure
 
 --- Chunk 2 ---
-Metadata: {header=Long Section Header, index=1, id=654}
+Metadata: {header=Long Section Header, index=1, id=525}
  irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
 
 Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt explicabo. Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut fugit,
 
 --- Chunk 3 ---
-Metadata: {header=Long Section Header, index=2, id=696}
+Metadata: {header=Long Section Header, index=2, id=567}
  fugit, sed quia consequuntur magni dolores eos qui ratione voluptatem sequi nesciunt. Neque porro quisquam est, qui dolorem ipsum quia dolor sit amet, consectetur, adipisci velit, sed quia non numquam eius modi tempora incidunt ut labore et dolore magnam aliquam quaerat voluptatem.
 
 --- Chunk 4 ---
-Metadata: {index=3, id=699}
+Metadata: {index=3, id=570}
  voluptatem.
 
 At vero eos et accusamus et iusto odio dignissimos ducimus qui blanditiis praesentium voluptatum deleniti atque corrupti quos dolores et quas molestias excepturi sint occaecati cupiditate non provident, similique sunt in culpa qui officia deserunt mollitia animi, id est laborum et dolorum fuga. Et harum quidem rerum facilis est et expedita distinctio.
@@ -30,7 +30,7 @@ At vero eos et accusamus et iusto odio dignissimos ducimus qui blanditiis praese
 
 
 --- Chunk 5 ---
-Metadata: {index=4, id=702}
+Metadata: {index=4, id=573}
 
 ### Subsection
 
@@ -40,7 +40,7 @@ This is a subsection under the long section. It also contains substantial conten
 
 
 --- Chunk 6 ---
-Metadata: {header=Another Section, index=5, id=704}
+Metadata: {header=Another Section, index=5, id=575}
 
 ## Another Section
