+++ native/src/test/resources/markdown-chunker-test/expected/sample5_500_0.txt
@@ -1,7 +1,7 @@
 500 0
 
 --- Chunk 1 ---
-Metadata: {index=0, id=4003}
+Metadata: {index=0, id=12}
 # Code Blocks and Horizontal Lines Test
 
 This file tests the markdown chunker's ability to handle code blocks and horizontal lines as delimiters.
@@ -23,7 +23,7 @@ After the code block, we continue with regular text. The chunker should split on
 
 
 --- Chunk 2 ---
-Metadata: {index=1, id=4017}
+Metadata: {index=1, id=16}
 
 ## Section After Asterisk Horizontal Line
 
@@ -44,7 +44,7 @@ More text after the Python code block.
 
 
 --- Chunk 3 ---
-Metadata: {index=2, id=4019}
+Metadata: {index=2, id=19}
 
 This section follows a horizontal line made with dashes. This tests the dash delimiter pattern.
 
@@ -64,7 +64,7 @@ ___
 
 
 --- Chunk 4 ---
-Metadata: {index=3, id=4021}
+Metadata: {index=3, id=23}
 
 This section follows a horizontal line made with underscores. Testing the underscore delimiter.
