+++ native/src/test/resources/markdown-chunker-test/expected/sample5_500_50.txt
@@ -1,7 +1,7 @@
 500 50
 
 --- Chunk 1 ---
-Metadata: {index=0, id=717}
+Metadata: {index=0, id=12}
 # Code Blocks and Horizontal Lines Test
 
 This file tests the markdown chunker's ability to handle code blocks and horizontal lines as delimiters.
@@ -23,7 +23,7 @@ After the code block, we continue with regular text. The chunker should split on
 
 
 --- Chunk 2 ---
-Metadata: {index=1, id=821}
+Metadata: {index=1, id=116}
 
 
 ***
@@ -47,7 +47,7 @@ More text after the Python code block.
 
 
 --- Chunk 3 ---
-Metadata: {index=2, id=824}
+Metadata: {index=2, id=119}
 
 ## Section After Dash Horizontal Line  
 
@@ -69,7 +69,7 @@ ___
 
 
 --- Chunk 4 ---
-Metadata: {index=3, id=827}
+Metadata: {index=3, id=122}
 
 ## Section After Underscore Horizontal Line
 
@@ -87,7 +87,7 @@ Final section with more content to ensure proper chunking behavior.
 
 
 --- Chunk 5 ---
-Metadata: {header=Multiple Code Blocks, index=4, id=829}
+Metadata: {header=Multiple Code Blocks, index=4, id=124}
 
 ## Multiple Code Blocks
