+++ native/src/test/resources/markdown-chunker-test/expected/sample6_500_0.txt
@@ -1,7 +1,7 @@
 500 0
 
 --- Chunk 1 ---
-Metadata: {index=0, id=4033}
+Metadata: {index=0, id=27}
 # Horizontal Line Variations Test
 
 This document specifically tests different horizontal line patterns to validate the splitter patterns.
@@ -22,7 +22,7 @@ This section comes after four asterisks. The pattern should match \n\*\*\*+\n wh
 
 
 --- Chunk 2 ---
-Metadata: {index=1, id=4066}
+Metadata: {index=1, id=40}
 
 Testing with five asterisks to ensure the + quantifier works correctly.
 
@@ -62,7 +62,7 @@ ____
 
 
 --- Chunk 3 ---
-Metadata: {index=2, id=4068}
+Metadata: {index=2, id=46}
 
 Testing with four underscores.
