+++ native/src/test/resources/markdown-chunker-test/expected/sample6_500_50.txt
@@ -1,7 +1,7 @@
 500 50
 
 --- Chunk 1 ---
-Metadata: {index=0, id=857}
+Metadata: {index=0, id=27}
 # Horizontal Line Variations Test
 
 This document specifically tests different horizontal line patterns to validate the splitter patterns.
@@ -22,7 +22,7 @@ This section comes after four asterisks. The pattern should match \n\*\*\*+\n wh
 
 
 --- Chunk 2 ---
-Metadata: {index=1, id=869}
+Metadata: {index=1, id=39}
 
 ## After Five Asterisks
 
@@ -62,7 +62,7 @@ ____
 
 
 --- Chunk 3 ---
-Metadata: {index=2, id=904}
+Metadata: {index=2, id=74}
 
 
 ____
