+++ native/src/test/resources/markdown-chunker-test/expected/sample7_500_50.txt
@@ -1,7 +1,7 @@
 500 50
 
 --- Chunk 1 ---
-Metadata: {index=0, id=945}
+Metadata: {index=0, id=40}
 # Code Block Test Document
 
 This document contains various code blocks to test the chunking functionality.
@@ -17,7 +17,7 @@ Here's a Java class with some methods:
 
 
 --- Chunk 2 ---
-Metadata: {header=Java Code Example, index=1, language=java, id=920, type=code_block}
+Metadata: {header=Java Code Example, index=1, language=java, id=15, type=code_block}
 ```java
 public class ExampleClass {
     private String name;
@@ -45,7 +45,7 @@ public class ExampleClass {
         return "ExampleClass{name='" + name + "', value=" + value + "}"
 
 --- Chunk 3 ---
-Metadata: {header=Java Code Example, index=2, language=java, id=921, type=code_block, prev=916}
+Metadata: {header=Java Code Example, index=2, language=java, id=16, type=code_block, prev=11}
 ;
     }
 }
@@ -53,7 +53,7 @@ Metadata: {header=Java Code Example, index=2, language=java, id=921, type=code_b
 
 
 --- Chunk 4 ---
-Metadata: {index=3, id=947}
+Metadata: {index=3, id=42}
 
 ## Python Code Example
 
@@ -81,7 +81,7 @@ if __name__ == "__main__":
 
 
 --- Chunk 5 ---
-Metadata: {index=4, id=951}
+Metadata: {index=4, id=46}
 
 ## Code Block Without Language
 
@@ -100,7 +100,7 @@ Here's some JavaScript code:
 
 
 --- Chunk 6 ---
-Metadata: {header=JavaScript Example, index=5, language=javascript, id=938, type=code_block}
+Metadata: {header=JavaScript Example, index=5, language=javascript, id=33, type=code_block}
 ```javascript
 class Calculator {
     constructor() {
@@ -129,14 +129,14 @@ console.log(calc.add(5, 3));
 console.log(calc.subtract(10, 4
 
 --- Chunk 7 ---
-Metadata: {header=JavaScript Example, index=6, language=javascript, id=939, type=code_block, prev=934}
+Metadata: {header=JavaScript Example, index=6, language=javascript, id=34, type=code_block, prev=29}
 ));
 console.log(calc.getHistory());
 ```
 
 
 --- Chunk 8 ---
-Metadata: {header=Conclusion, index=7, id=952}
+Metadata: {header=Conclusion, index=7, id=47}
 
 ## Conclusion
