+++ native/src/test/resources/markdown-chunker-test/expected/sample8_500_0.txt
@@ -1,7 +1,7 @@
 500 0
 
 --- Chunk 1 ---
-Metadata: {index=0, id=4122}
+Metadata: {index=0, id=5}
 This is a simple document without headers to test code block chunking.
 
 Here's a Java code block:
@@ -9,7 +9,7 @@ Here's a Java code block:
 
 
 --- Chunk 2 ---
-Metadata: {index=1, language=java, id=4123, type=code_block}
+Metadata: {index=1, language=java, id=6, type=code_block}
 ```java
 public class SimpleClass {
     private String name;
@@ -30,14 +30,14 @@ public class SimpleClass {
 
 
 --- Chunk 3 ---
-Metadata: {index=2, id=4124}
+Metadata: {index=2, id=7}
 
 And here's a Python code block:
 
 
 
 --- Chunk 4 ---
-Metadata: {index=3, language=python, id=4125, type=code_block}
+Metadata: {index=3, language=python, id=8, type=code_block}
 ```python
 def hello_world():
     print("Hello, World!")
@@ -52,14 +52,14 @@ if __name__ == "__main__":
 
 
 --- Chunk 5 ---
-Metadata: {index=4, id=4126}
+Metadata: {index=4, id=9}
 
 And a code block without language:
 
 
 
 --- Chunk 6 ---
-Metadata: {index=5, language=unknown, id=4127, type=code_block}
+Metadata: {index=5, language=unknown, id=10, type=code_block}
 ```
 This is a code block without language specification.
 It should still be treated as a code block.
@@ -67,6 +67,6 @@ It should still be treated as a code block.
 
 
 --- Chunk 7 ---
-Metadata: {index=6, id=4128}
+Metadata: {index=6, id=11}
 
 This is the end of the document.
